Rig suit glove/boots and tweaks

+++ code/modules/clothing/spacesuits/rig.dm
@@ -97,27 +97,45 @@
 	allowed = list(/obj/item/device/flashlight,/obj/item/weapon/tank,/obj/item/weapon/storage/bag/ore,/obj/item/device/t_scanner,/obj/item/weapon/pickaxe, /obj/item/device/rcd, /obj/item/weapon/wrench/socket)
 	max_heat_protection_temperature = SPACE_SUIT_MAX_HEAT_PROTECTION_TEMPERATURE
 	pressure_resistance = 200 * ONE_ATMOSPHERE
-	var/obj/item/clothing/head/helmet/space/rig/H = null
-	var/head_type = /obj/item/clothing/head/helmet/space/rig
-	var/obj/item/weapon/cell/cell = null
-	var/cell_type = /obj/item/weapon/cell/high //The cell_type we're actually using
+	var/activated = FALSE
 	var/list/modules = list()
-	actions_types = list(/datum/action/item_action/toggle_rig_helmet)
+	actions_types = list(/datum/action/item_action/toggle_rig_suit)
+
+	var/obj/item/clothing/head/helmet/space/rig/H = null
+	var/obj/item/clothing/gloves/G = null
+	var/obj/item/clothing/shoes/magboots/MB = null
+	var/obj/item/weapon/cell/cell = null
+
+	var/head_type = /obj/item/clothing/head/helmet/space/rig
+	var/boots_type =  null
+	var/gloves_type = null
+	var/cell_type = /obj/item/weapon/cell/high //The cell_type we're actually using
 
 /obj/item/clothing/suit/space/rig/New()
 	..()
-	cell = new cell_type
-	H = new head_type
+	if(cell_type)
+		cell = new cell_type(src)
+	if(head_type)
+		H = new head_type(src)
+		H.rig = src
+	if(gloves_type)
+		G = new gloves_type(src)
+	if(boots_type)
+		MB = new boots_type(src)
 
 /obj/item/clothing/suit/space/rig/Destroy()
-	qdel(cell)
-	cell = null
-	if(H && (H.loc == src || !H.loc))
-		qdel(H)
+	for(var/obj/item/clothing/C in list(H,G,MB))
+		if(C && (C.loc == src || !C.loc))
+			qdel(C)
 	H = null
+	G = null
+	MB = null
 	for(var/obj/M in modules)
 		qdel(M)
 	modules.Cut()
+	if(cell)
+		qdel(cell)
+	cell = null
 	..()
 
 /obj/item/clothing/suit/space/rig/examine(mob/user)
@@ -125,46 +143,92 @@
 	for(var/obj/item/rig_module/M in modules)
 		M.examine_addition(user)
 
+/obj/item/clothing/suit/space/rig/get_cell()
+	return cell
+
 /obj/item/clothing/suit/space/rig/unequipped(mob/living/carbon/human/user, var/from_slot = null)
 	..()
 	if(from_slot == slot_wear_suit && istype(user))
 		deactivate_suit(user)
 
-/obj/item/clothing/suit/space/rig/proc/toggle_helmet(mob/living/carbon/human/user)
+/obj/item/clothing/suit/space/rig/proc/toggle_suit(mob/living/carbon/human/user)
 	if(!user.is_wearing_item(src, slot_wear_suit))
 		return
+	!activated ? initialize_suit(user) : deactivate_suit(user)
+
+/obj/item/clothing/suit/space/rig/proc/initialize_suit(mob/living/carbon/human/user)
 	if(H)
-		if(!user.head)
-			to_chat(user, "<span class = 'notice'>\The [H] extends from \the [src].</span>")
-			user.equip_to_slot(H, slot_head)
-			H.rig = src
-			H = null
-			initialize_suit(user)
-	else
-		if(user.head && istype(user.head, head_type))
-			var/obj/I = user.head
-			to_chat(user, "<span class = 'notice'>\The [I] retracts into \the [src].</span>")
-			user.u_equip(I,0)
-			I.forceMove(src)
-			H = I
-			deactivate_suit(user)
-
-
-/obj/item/clothing/suit/space/rig/proc/initialize_suit(mob/user)
+		if(user.head)
+			user.remove_from_mob(user.head)
+		to_chat(user, "<span class = 'notice'>\The [H] extends from \the [src].</span>")
+		user.equip_to_slot(H, slot_head)
+		H = null
+	if(G)
+		if(user.gloves)
+			user.remove_from_mob(user.gloves)
+		to_chat(user, "<span class = 'notice'>\The [G] extends from \the [src].</span>")
+		user.equip_to_slot(G, slot_gloves)
+		G = null
+	if(MB)
+		if(user.shoes)
+			user.remove_from_mob(user.shoes)
+		to_chat(user, "<span class = 'notice'>\The [MB] extends from \the [src].</span>")
+		user.equip_to_slot(MB, slot_shoes)
+		MB = null
 	for(var/obj/item/rig_module/R in modules)
 		R.activate(user,src)
+	activated = TRUE
 
-/obj/item/clothing/suit/space/rig/proc/deactivate_suit(mob/user)
+/obj/item/clothing/suit/space/rig/proc/deactivate_suit(mob/living/carbon/human/user)
+	if(head_type && user.head)
+		if(!H)
+			if(istype(user.head, head_type))
+				var/obj/item/clothing/UH = user.head
+				to_chat(user, "<span class = 'notice'>\The [UH] retracts into \the [src].</span>")
+				user.u_equip(UH,0)
+				UH.forceMove(src)
+				H = UH
+			else
+				to_chat(user, "<span class = 'warning'>\The [user.head] isn't compatible with \the [src].</span>")
+	if(gloves_type && user.gloves)
+		if(!G)
+			if(istype(user.gloves, gloves_type))
+				var/obj/item/clothing/UG = user.gloves
+				to_chat(user, "<span class = 'notice'>\The [UG] retracts into \the [src].</span>")
+				user.u_equip(UG,0)
+				UG.forceMove(src)
+				G = UG
+			else
+				to_chat(user, "<span class = 'warning'>\The [user.gloves] isn't compatible with \the [src].</span>")
+	if(boots_type && user.shoes)
+		if(!MB)
+			if(istype(user.shoes, boots_type))
+				var/obj/item/clothing/UMB = user.shoes
+				to_chat(user, "<span class = 'notice'>\The [UMB] retracts into \the [src].</span>")
+				user.u_equip(UMB,0)
+				UMB.forceMove(src)
+				MB = UMB
+			else
+				to_chat(user, "<span class = 'warning'>\The [user.shoes] isn't compatible with \the [src].</span>")
 	for(var/obj/item/rig_module/R in modules)
 		R.deactivate(user,src)
+	activated = FALSE
 
 /obj/item/clothing/suit/space/rig/attackby(obj/W, mob/user)
-	if(!H && istype(W, head_type) && user.drop_item(W, src, force_drop = 1))
+	if(!H && istype(W, head_type) && user.drop_item(W, src, force_drop = TRUE))
 		to_chat(user, "<span class = 'notice'>You attach \the [W] to \the [src].</span>")
 		H = W
+		H.rig = src
+		return
+	if(!G && istype(W, gloves_type) && user.drop_item(W, src, force_drop = TRUE))
+		to_chat(user, "<span class = 'notice'>You attach \the [W] to \the [src].</span>")
+		G = W
+		return
+	if(!MB && istype(W, boots_type) && user.drop_item(W, src, force_drop = TRUE))
+		to_chat(user, "<span class = 'notice'>You attach \the [W] to \the [src].</span>")
+		MB = W
 		return
 	..()
-
 
 //Chief Engineer's rig
 /obj/item/clothing/head/helmet/space/rig/elite
@@ -189,6 +253,11 @@
 	clothing_flags = PLASMAGUARD
 	cell_type = /obj/item/weapon/cell/super
 	head_type = /obj/item/clothing/head/helmet/space/rig/elite
+
+/obj/item/clothing/suit/space/rig/elite/test
+	name = "testing-type advanced hardsuit"
+	gloves_type = /obj/item/clothing/gloves/yellow
+	boots_type = /obj/item/clothing/shoes/magboots/elite
 
 //Mining rig
 /obj/item/clothing/head/helmet/space/rig/mining
